cambios en los mensajes de error y bloquero de cuenta, enviar mail cuando se cancela la clase

+++ gestionDePiletaSportClub/Controllers/Api/ActivitiesController.cs
@@ -10,6 +10,8 @@ using gestionDePiletaSportClub.ViewModels;
 using gestionDePiletaSportClub.Dtos;
 using System.Data.Entity;
 using AutoMapper;
+using System.Threading.Tasks;
+using Microsoft.AspNet.Identity;
 
 namespace gestionDePiletaSportClub.Controllers.Api
 {
@@ -42,7 +44,11 @@ namespace gestionDePiletaSportClub.Controllers.Api
         //GET /api/activities/{id}
         public EventDto GetActivity(int Id)
         {
-            var activity = _context.Actividad.Include(a => a.TipoActividad).SingleOrDefault(a => a.Id == Id);
+            var activity = _context.Actividad
+                .Include(a => a.TipoActividad)
+                .Include(a => a.Level)
+                .Include(a => a.MembershipType)
+                .SingleOrDefault(a => a.Id == Id);
 
             return new EventDto(Mapper.Map<Actividad, ActivityDto>(activity));
         }
@@ -50,25 +56,41 @@ namespace gestionDePiletaSportClub.Controllers.Api
         
         [HttpPut]
         [Route("api/activities/{Id}/cancelar")]
-        public void UpdateUser(int Id)
+        public async void CancelActivity(int Id)
         {
-            var activity = _context.Actividad.SingleOrDefault(a => a.Id == Id);
+            var activity = _context.Actividad
+                .Include(a=>a.TipoActividad)
+                .Include(a=>a.Level)
+                .Include(a=>a.MembershipType)
+                .SingleOrDefault(a => a.Id == Id);
 
             activity.EstadoActividadId = EstadoActividad.Cancelada;
 
             var enrollments = _context.Enrollment.Where(e => e.ActividadId == activity.Id).Include(e=>e.ApplicationUser).ToList();
-
+            List<string> emailList = new List<string>();
             foreach (Enrollment e in enrollments) {
                 e.ApplicationUser.AmountOfPendingActivities++;
-                //AVISAR QUE LA CLASE FUE CANCELADA
-                
+                emailList.Add(e.ApplicationUser.Email);
             }
-            _context.Enrollment.RemoveRange(enrollments);
-
+            if (enrollments.Count > 0)
+            {
+                string emails = string.Join(",", emailList.ToArray());
+                 InformarCancelarActividad(emails, activity);
+                _context.Enrollment.RemoveRange(enrollments);
+            }
             _context.SaveChanges();
         }
 
-
-
+        private void InformarCancelarActividad(string email, Actividad activity)
+        {
+            EmailService e = new EmailService();
+            IdentityMessage message = new IdentityMessage();
+            message.Subject = "[Sport Club] Clase " + activity.Schedule.ToString()+" cancelada ";
+            message.Body = String.Format("La clase de {0} para el nivel {1} de {2} ha sido cancelada. La misma no sera descontada de su plan de actividades.<br/> " +
+                "Disculpe las molestias ocasionadas <br/>" +
+                "<strong>Equipo de Sport Club </strong>",activity.TipoActividad.Name, activity.Level.Name,activity.MembershipType.Name);
+            message.Destination = email;
+            e.Send(message);
+        }
     }
 }
